client-or-server-contained-multi-properties-itest: New Pure Maven Profile Activation
Added another example which is especially meant to get an idea how profile activation
influences the order of activated profiles. In the new example it does not matter if
you specify `-Daut-client -Daut-server`, or `-Daut-server -Daut-client`: Both will
result in the `aut-server` profile to be executed last.

This may be important, if you want to set also other properties than unmatchedAut if
a certain AUT is specified.